adapt for arduino ide 1.8.2
in my context (Arduino IDE 1.8.2, Mac, Adafruit Feather 32u4 with Lora) the C-Compiler is doing late type conversion from int to float when combining the lower and the higher byte. To fix this I put '.0' behin 256 and 10 so they start out as floats. This fixes the problem and should also work with other platforms.